Better behaviour of the activity gauge in the docker. Goal for the victim: keep that gauge at zero as much as possible. Not really productive, but health++.
svn path=/trunk/playground/utils/rsibreak/; revision=487571

+++ src/rsitimer.h
@@ -60,7 +60,7 @@ class RSITimer : public QObject
     signals:
         void breakNow();
         void setCounters( const QTime & );
-        void updateIdleAvg( int );
+        void updateIdleAvg( double );
         void minimize();
         void relax( int );
