+++ src/rsiwidget.h
@@ -62,7 +62,7 @@ class RSIWidget : public QWidget
         void minimize();
         void maximize();
         void setCounters( const QTime & );
-        void updateIdleAvg( int );
+        void updateIdleAvg( double );
         void readConfig();
 
     protected:
